update tabs for stream profile
