+++ src/core/client/stream/tabs/Profile/Preferences/PreferencesContainer.tsx
@@ -1,13 +1,12 @@
 import React, { FunctionComponent } from "react";
 
 import { graphql, withFragmentContainer } from "coral-framework/lib/relay";
-import {} from "coral-ui/components/v2";
+import { HorizontalGutter } from "coral-ui/components/v2";
 
 import IgnoreUserSettingsContainer from "./IgnoreUserSettingsContainer";
 import NotificationSettingsContainer from "./NotificationSettingsContainer";
 
 import { PreferencesContainer_viewer } from "coral-stream/__generated__/PreferencesContainer_viewer.graphql";
-// import styles from "./PreferencesContainer.css";
 
 interface Props {
   viewer: PreferencesContainer_viewer;
@@ -15,10 +14,10 @@ interface Props {
 
 const PreferencesContainer: FunctionComponent<Props> = props => {
   return (
-    <div>
+    <HorizontalGutter spacing={4}>
       <NotificationSettingsContainer viewer={props.viewer} />
       <IgnoreUserSettingsContainer viewer={props.viewer} />
-    </div>
+    </HorizontalGutter>
   );
 };
