+++ src/core/client/stream/test/profile/account.spec.tsx
@@ -7,12 +7,10 @@ import {
   createResolversStub,
   CreateTestRendererParams,
   waitForElement,
-  waitUntilThrow,
   within,
 } from "coral-framework/testHelpers";
 
 import {
-  commenters,
   settings,
   settingsWithoutLocalAuth,
   stories,
@@ -46,10 +44,6 @@ async function createTestRenderer(
     },
   });
 
-  const ignoredCommenters = await waitForElement(() =>
-    within(testRenderer.root).queryByTestID("profile-account-ignoredCommenters")
-  );
-
   const changePassword = within(testRenderer.root).queryByTestID(
     "profile-account-changePassword"
   );
@@ -57,7 +51,6 @@ async function createTestRenderer(
   return {
     testRenderer,
     context,
-    ignoredCommenters,
     changePassword,
   };
 }
@@ -67,6 +60,7 @@ it("renders the empty settings pane", async () => {
     testRenderer: { root },
   } = await createTestRenderer();
   expect(within(root).toJSON()).toMatchSnapshot();
+  within(root).debug();
   expect(await within(root).axe()).toHaveNoViolations();
 });
 
@@ -142,58 +136,4 @@ it("render password change form", async () => {
   });
 
   expect(updatePassword.calledOnce).toBeTruthy();
-});
-
-it("render empty ignored users list", async () => {
-  const { ignoredCommenters } = await createTestRenderer();
-  const editButton = within(ignoredCommenters).getByText("Manage");
-  act(() => {
-    editButton.props.onClick();
-  });
-  await waitForElement(() =>
-    within(ignoredCommenters).getByText(
-      "You are not currently ignoring anyone",
-      {
-        exact: false,
-      }
-    )
-  );
-});
-
-it("render ignored users list", async () => {
-  const { ignoredCommenters } = await createTestRenderer({
-    resolvers: createResolversStub<GQLResolver>({
-      Query: {
-        viewer: () =>
-          pureMerge<typeof viewer>(viewer, {
-            ignoredUsers: [commenters[0], commenters[1]],
-          }),
-      },
-      Mutation: {
-        removeUserIgnore: ({ variables }) => {
-          expectAndFail(variables).toMatchObject({
-            userID: commenters[0].id,
-          });
-          return {};
-        },
-      },
-    }),
-  });
-  const editButton = within(ignoredCommenters).getByText("Manage");
-  act(() => {
-    editButton.props.onClick();
-  });
-  within(ignoredCommenters).getByText(commenters[0].username!);
-  within(ignoredCommenters).getByText(commenters[1].username!);
-
-  // Stop ignoring first users.
-  within(ignoredCommenters)
-    .getAllByText("Stop ignoring", { selector: "button" })[0]
-    .props.onClick();
-
-  // First user should dissappear from list.
-  await waitUntilThrow(() =>
-    within(ignoredCommenters).getByText(commenters[0].username!)
-  );
-  within(ignoredCommenters).getByText(commenters[1].username!);
 });
